+++ docs/changelog.rst
@@ -7,3 +7,19 @@ v0.1.0
 ------
 
 Initial release of the Qt version.
+
+
+v0.2.0 (unreleased)
+-------------------
+
+Added:
+^^^^^^
+
+* A global ``font`` style option to set all fonts at once. If a local option such as
+  ``statusbar.font`` is defined, it overrides the global option.
+
+Changed:
+^^^^^^^^
+
+* All styles are now based upon base16. Therefore custom styles must define the colors
+  ``base00`` to ``base0f``. All other style options are optional.